refactor(agent): isolate pi chat follow-up state
